feat: implement Scheduled Maintenance section and refactor related components
